Change int type to double

+++ src/main/java/sample/Boid.java
@@ -2,27 +2,27 @@ package sample;
 
 public abstract class Boid {
 
-    private int[] position;// 0->X; 1->Y
-    private int[] velocity;// 0->VX; 1->VY
+    private double[] position;// 0->X; 1->Y
+    private double[] velocity;// 0->VX; 1->VY
 
-    public Boid(int[] position, int[] velocity) {
+    public Boid(double[] position, double[] velocity) {
         this.position = position;
         this.velocity = velocity;
     }
 
-    public int[] getPosition() {
+    public double[] getPosition() {
         return position;
     }
 
-    public void setPosition(int[] position) {
+    public void setPosition(double[] position) {
         this.position = position;
     }
 
-    public int[] getVelocity() {
+    public double[] getVelocity() {
         return velocity;
     }
 
-    public void setVelocity(int[] velocity) {
+    public void setVelocity(double[] velocity) {
         this.velocity = velocity;
     }
 }
